+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Calendar Versioning](https://calver.org).
 
 ### Added
 - [CSE] Added tests for deletion notifications (*creator* attribute).
+- [CSE] Added support for xs:token type in attribute policies and validation.
 
 ### Experimental